requirements.txt
python-telegram-bot==13.15
redis==3.2.1
environs==10.3.0
